# seccion-2-spec-driven-back/openapi.yaml
openapi: 3.0.3
info:
  title: Cotizador de Reparación de Calzado
  description: |
    API REST para generar cotizaciones de reparación de calzado.
    
    Expone tres grupos de endpoints:
    1. **Catálogo de Tipos de Calzado** — consultar los tipos de calzado disponibles.
    2. **Catálogo de Reparaciones** — consultar las reparaciones disponibles.
    3. **Cotizaciones** — generar una cotización estimada (costo y tiempo de entrega).
    
    Todos los datos se almacenan en memoria. No requiere autenticación.
  version: 1.0.0
  contact:
    name: Shoe Repair Quotation API

servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo local

tags:
  - name: Catálogo
    description: Endpoints para consultar los catálogos de tipos de calzado y reparaciones
  - name: Cotizaciones
    description: Endpoint para generar cotizaciones de reparación

paths:
  /catalog/shoe-types:
    get:
      tags:
        - Catálogo
      summary: Listar tipos de calzado
      description: |
        Devuelve todos los tipos de calzado disponibles en el catálogo.
        Retorna una lista vacía `[]` si no hay registros; nunca `null`.
      operationId: getAllShoeTypes
      responses:
        "200":
          description: Lista de tipos de calzado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShoeType"
              example:
                - id: "st-1"
                  name: "Zapatilla deportiva"
                  complexityFactor: 1.0
                - id: "st-2"
                  name: "Bota"
                  complexityFactor: 1.5
                - id: "st-3"
                  name: "Sandalia"
                  complexityFactor: 0.8

  /catalog/repairs:
    get:
      tags:
        - Catálogo
      summary: Listar reparaciones disponibles
      description: |
        Devuelve todas las reparaciones disponibles en el catálogo.
        Retorna una lista vacía `[]` si no hay registros; nunca `null`.
      operationId: getAllRepairs
      responses:
        "200":
          description: Lista de reparaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Repair"
              example:
                - id: "r-1"
                  name: "Cambio de suela"
                  basePrice: 25.00
                  estimatedDays: 5
                - id: "r-2"
                  name: "Costura"
                  basePrice: 15.00
                  estimatedDays: 3
                - id: "r-3"
                  name: "Limpieza profunda"
                  basePrice: 10.00
                  estimatedDays: 1
                - id: "r-4"
                  name: "Cambio de taco"
                  basePrice: 20.00
                  estimatedDays: 4

  /quotations:
    post:
      tags:
        - Cotizaciones
      summary: Generar una cotización
      description: |
        Genera una cotización estimada de costo y tiempo de entrega a partir de un tipo de calzado y una o más reparaciones.

        **Reglas de negocio:**
        - `subtotal = Σ (repair.basePrice × shoeType.complexityFactor)`
        - Si `urgent = true`: `surcharge = subtotal × 0.30`, `total = subtotal + surcharge`
        - Si `urgent = false`: `surcharge = 0`, `total = subtotal`
        - `estimatedDays` sin urgencia: `max(repair.estimatedDays)`
        - `estimatedDays` con urgencia: `max(1, ceil(max(repair.estimatedDays) / 2))`

        Todos los valores monetarios se redondean a 2 decimales (half-up).
      operationId: createQuotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuotationRequest"
            examples:
              sinUrgencia:
                summary: Cotización sin urgencia
                value:
                  shoeTypeId: "st-1"
                  repairIds: ["r-1", "r-2"]
                  urgent: false
              conUrgencia:
                summary: Cotización con urgencia
                value:
                  shoeTypeId: "st-2"
                  repairIds: ["r-1", "r-3"]
                  urgent: true
      responses:
        "201":
          description: Cotización generada exitosamente
          content:
            application/json:
              schema:
                type: object
                required:
                  - quotation
                properties:
                  quotation:
                    $ref: "#/components/schemas/Quotation"
              examples:
                sinUrgencia:
                  summary: Respuesta sin urgencia
                  value:
                    quotation:
                      id: "550e8400-e29b-41d4-a716-446655440000"
                      createdAt: "2024-06-15T10:30:00.000Z"
                      shoeTypeId: "st-1"
                      repairIds: ["r-1", "r-2"]
                      subtotal: 40.00
                      surcharge: 0
                      total: 40.00
                      estimatedDays: 5
                      urgent: false
                conUrgencia:
                  summary: Respuesta con urgencia
                  value:
                    quotation:
                      id: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                      createdAt: "2024-06-15T10:31:00.000Z"
                      shoeTypeId: "st-2"
                      repairIds: ["r-1", "r-3"]
                      subtotal: 52.50
                      surcharge: 15.75
                      total: 68.25
                      estimatedDays: 3
                      urgent: true
        "400":
          description: |
            Error de validación. Causas posibles:
            - `repairIds` vacío o ausente
            - `shoeTypeId` no es un string válido
            - `urgent` no es un booleano
            - `shoeTypeId` no existe en el catálogo
            - Uno o más `repairIds` no existen en el catálogo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                camposRequeridos:
                  summary: Campos faltantes o inválidos
                  value:
                    error: "Invalid request body"
                    details:
                      - "shoeTypeId must be a non-empty string"
                      - "repairIds must be a non-empty array"
                      - "urgent must be a boolean"
                tipoCalzadoNoEncontrado:
                  summary: Tipo de calzado no encontrado
                  value:
                    error: 'Shoe type "st-unknown" not found'
                    details:
                      - "SHOE_TYPE_NOT_FOUND"
                reparacionesNoEncontradas:
                  summary: Reparaciones no encontradas
                  value:
                    error: "Repairs not found: r-999, r-888"
                    details:
                      - "REPAIRS_NOT_FOUND"
                repairIdsVacio:
                  summary: Lista de reparaciones vacía
                  value:
                    error: "Invalid request body"
                    details:
                      - "repairIds must be a non-empty array"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Error interno del servidor"

components:
  schemas:
    ShoeType:
      type: object
      description: Tipo de calzado registrado en el catálogo
      required:
        - id
        - name
        - complexityFactor
      properties:
        id:
          type: string
          description: Identificador único del tipo de calzado
          example: "st-1"
        name:
          type: string
          description: Nombre legible del tipo de calzado
          example: "Zapatilla deportiva"
        complexityFactor:
          type: number
          format: double
          description: |
            Multiplicador aplicado al precio base de cada reparación.
            Siempre mayor que 0.
          minimum: 0.01
          example: 1.0

    Repair:
      type: object
      description: Reparación registrada en el catálogo
      required:
        - id
        - name
        - basePrice
        - estimatedDays
      properties:
        id:
          type: string
          description: Identificador único de la reparación
          example: "r-1"
        name:
          type: string
          description: Nombre legible de la reparación
          example: "Cambio de suela"
        basePrice:
          type: number
          format: double
          description: |
            Precio base de la reparación en unidades monetarias locales.
            Siempre mayor que 0.
          minimum: 0.01
          example: 25.00
        estimatedDays:
          type: integer
          format: int32
          description: |
            Número estimado de días hábiles para completar la reparación.
            Siempre mayor o igual a 1.
          minimum: 1
          example: 5

    QuotationRequest:
      type: object
      description: Cuerpo de la solicitud para generar una cotización
      required:
        - shoeTypeId
        - repairIds
        - urgent
      properties:
        shoeTypeId:
          type: string
          description: |
            ID del tipo de calzado. Debe existir en el catálogo.
          example: "st-1"
        repairIds:
          type: array
          description: |
            IDs de las reparaciones a incluir. Debe contener al menos 1 elemento.
            Todos los IDs deben existir en el catálogo.
          items:
            type: string
          minItems: 1
          example: ["r-1", "r-2"]
        urgent:
          type: boolean
          description: |
            Indica si el cliente solicita procesamiento urgente.
            - `true`: se aplica recargo del 30% y el tiempo de entrega se reduce a la mitad (mínimo 1 día).
            - `false`: sin recargo, tiempo estándar.
          example: false

    Quotation:
      type: object
      description: Cotización generada por el sistema
      required:
        - id
        - createdAt
        - shoeTypeId
        - repairIds
        - subtotal
        - surcharge
        - total
        - estimatedDays
        - urgent
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único de la cotización (UUID v4)
          example: "550e8400-e29b-41d4-a716-446655440000"
        createdAt:
          type: string
          format: date-time
          description: Fecha y hora de creación en formato ISO 8601
          example: "2024-06-15T10:30:00.000Z"
        shoeTypeId:
          type: string
          description: ID del tipo de calzado utilizado
          example: "st-1"
        repairIds:
          type: array
          description: IDs de las reparaciones incluidas
          items:
            type: string
          minItems: 1
          example: ["r-1", "r-2"]
        subtotal:
          type: number
          format: double
          description: |
            Suma de (repair.basePrice × shoeType.complexityFactor) para cada reparación.
            Redondeado a 2 decimales.
          example: 40.00
        surcharge:
          type: number
          format: double
          description: |
            Recargo por urgencia.
            - Si `urgent = true`: `subtotal × 0.30`
            - Si `urgent = false`: `0`
          example: 0
        total:
          type: number
          format: double
          description: |
            Precio total = subtotal + surcharge.
            Redondeado a 2 decimales.
          example: 40.00
        estimatedDays:
          type: integer
          format: int32
          description: |
            Días estimados de entrega.
            - Sin urgencia: `max(repair.estimatedDays)`
            - Con urgencia: `max(1, ceil(max(repair.estimatedDays) / 2))`
            Siempre un entero ≥ 1.
          minimum: 1
          example: 5
        urgent:
          type: boolean
          description: Indica si la cotización fue solicitada como urgente
          example: false

    ErrorResponse:
      type: object
      description: Formato estándar de respuesta de error
      required:
        - error
      properties:
        error:
          type: string
          description: Descripción principal del error
          example: "Invalid request body"
        details:
          type: array
          description: |
            Lista opcional de detalles adicionales.
            Se usa cuando hay múltiples problemas a reportar.
          items:
            type: string
          example: ["shoeTypeId must be a non-empty string"]